// src/schema.d.ts
export interface components {
// ...
    parameters: never;
// ...
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    "basic.getCharge": {
        parameters: {
            query: {
                /** @description The currency that the `price` is specified in. */
                currency: string;
                /**
                 * @description The price of the goods being sold in this transaction, in
                 *     the `currency`'s smallest unit. For example, if a trading
                 *     card is being sold for $12.34 (with `currency` as `usd`),
                 *     then the request for the charge for this transaction would
                 *     be `/charge?price=1234&currency=usd`.
                 */
                price: number;
                /**
                 * @description When creating transactions from a multi-use listing, the
                 *     `quantity` parameter can be provided in order to generate a
                 *     charge for the given price multiplied by the given quantity.
                 *     See `/multi_use_listings/{listingId}/create_transaction`
                 *     for more details.
                 */
                quantity?: number;
                /**
                 * @description The payment method that will be used to pay for the
                 *     transaction. This is necessary because different payment
                 *     methods may result in different fees.
                 *
                 *     The default value is `card`.
                 */
                payment_method?: string;
                use_hr_post?: boolean;
                /** @description The custom `postage_fee` for the transaction. */
                postage_fee?: number;
                /** @description The charge config for which the charge amount is computed. */
                charge_config?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["basic.Charge"];
                };
            };
            /**
             * @description Bad Request
             *
             *     `code` can be one of the following:
             *
             *       * `currency_missing`
             *       * `invalid_price`
             *       * `negative_price`
             *       * `price_too_low`
             *       * `unsupported_currency`
             *       * `invalid_charge_config`
             *       * `unsupported_combination`: The provided payment method, charge config and currency combination isn't supported.
             */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["Error"];
                };
            };
        };
    };
// ...
    "basic.createTransaction": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": {
                    /**
                     * Format: int64
                     * @description The `charge` value returned from a request to
                     *     `/charge`.
                     */
                    charge: number;
                    /**
                     * Format: int64
                     * @description The `charge_calculator_version` value returned from a request to `/charge`.
                     */
                    charge_calculator_version: number;
                    /**
                     * Format: int64
                     * @description The charge config for which the charge amount is computed.
                     */
                    charge_config?: number;
                    /**
                     * Format: int64
                     * @description The seller `charge` value returned from a
                     *     request to `/charge`.
                     */
                    charge_seller?: number;
                    client_id?: string;
                    /** @description The currency that the `price` is specified in. */
                    currency: string;
                    /** @description A description of the goods being sold. */
                    description: string;
                    /**
                     * @description `features` contains flags that modify the
                     *     transaction flow.
                     */
                    features?: components["schemas"]["basic.Feature"][];
                    /** @description The payment method to use for the transaction. */
                    payment_method?: string;
                    /**
                     * Format: int64
                     * @description The price of the goods being sold, in the
                     *     `currency`'s smallest unit.  The `charge` value
                     *     should correspond to the Trustap charge created
                     *     with this price, otherwise this request will
                     *     fail with a `400` error.
                     */
                    price: number;
                    role: components["schemas"]["basic.Role"];
                };
            };
        };
        responses: {
            /** @description Created */
            201: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["basic.Transaction"];
                };
            };
            /**
             * @description Bad Request
             *
             *     `code` can be one of the following:
             *
             *       * `incorrect_calc_version`
             *       * `incorrect_charge`
             *       * `incorrect_charge_seller`
             *       * `invalid_calc_version`
             *       * `negative_price`
             *       * `price_too_low`
             *       * `invalid_charge_config`
             */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["Error"];
                };
            };
        };
    };
// ...
}